last commit

+++ app/src/main/java/com/example/final_app_fmav/MainActivity.java
@@ -1,7 +1,6 @@
 package com.example.final_app_fmav;
 
 import android.content.Intent;
-import android.media.MediaPlayer;
 import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
@@ -16,11 +15,10 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.final_app_fmav.Models.Vehicle;
 import com.example.final_app_fmav.Utilities.DataManager;
+import com.example.final_app_fmav.Utilities.SingleGen;
 import com.google.android.material.textview.MaterialTextView;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
 
 import java.util.ArrayList;
 
@@ -29,15 +27,10 @@ public class MainActivity extends AppCompatActivity {
     private Button sign_In_BTN,sign_Out_BTN,yad2_BTN;
     private ImageView slogan_IMG;
     private SwitchCompat music_switch;
-    private MediaPlayer mediaPlayer;
-    // Write a message to the database
-    FirebaseDatabase database;
-    DatabaseReference myRef;
     public FirebaseAuth mAuth;
     public FirebaseUser currentUser;
     private ArrayList<Vehicle> vehicles;
     private MaterialTextView counter;
-    private boolean musicChecked;
 
 
 
@@ -49,11 +42,6 @@ public class MainActivity extends AppCompatActivity {
         mAuth = FirebaseAuth.getInstance();
         currentUser = mAuth.getCurrentUser();
         vehicles=DataManager.getVehicles();
-        //starting background sound
-        mediaPlayer = MediaPlayer.create(this,R.raw.drive);
-        mediaPlayer.setVolume(1f,1f);
-        mediaPlayer.setLooping(true);
-       // mediaPlayer.start();
 
 
         // init buttons and views
@@ -69,20 +57,18 @@ public class MainActivity extends AppCompatActivity {
         });
 
         //music on or off switch
-        Intent myIntent = new Intent(this, LogInActivity.class);
-        music_switch.setChecked(true);
+
         music_switch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(CompoundButton compoundButton, boolean b) {
                 if(b) {
-                    mediaPlayer.start();
-                    myIntent.putExtra(VehiclesLST.KEY_MUSIC, music_switch.isChecked());
+                    SingleGen.getInstance().mediaPlayerOn();
                 }
                 else
-                    mediaPlayer.pause();
-                    myIntent.putExtra(VehiclesLST.KEY_MUSIC, music_switch.isChecked());
+                    SingleGen.getInstance().mediaPlayerOff();
             }
         });
+
                 ////Sign Out User
         sign_Out_BTN.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -94,16 +80,17 @@ public class MainActivity extends AppCompatActivity {
         sign_In_BTN.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                openLogInWindow(myIntent);
+                openLogInWindow();
             }
         });
     }
 
     private void logout() {
-
-       currentUser.delete();
-       mAuth.signOut();
-       Toast.makeText(this,"Log Out Completed Successfully",Toast.LENGTH_SHORT).show();
+       if(currentUser!=null) {
+           currentUser.delete();
+           mAuth.signOut();
+           Toast.makeText(this, "Log Out Completed Successfully", Toast.LENGTH_SHORT).show();
+       }
     }
 
     private void findViews() {
@@ -116,43 +103,38 @@ public class MainActivity extends AppCompatActivity {
         counter.setText("There Are "+vehicles.size()+" "+counter.getText());
 
     }
-    private void openLogInWindow(Intent myIntent) {
-
+    private void openLogInWindow() {
+        Intent myIntent=new Intent(this,LogInActivity .class);
         startActivity(myIntent);
 
     }
-    private void openVehiclesLSTWindow() {
-        Intent slogan = new Intent(this,VehiclesLST.class);
-        startActivity(slogan);
-    }
-    private void openSignUpWindow() {
-       // Intent signIn = new Intent(this,SignUpActivity.class);
-       // startActivity(signIn);
-    }
+
     @Override
     protected void onPause() {
         super.onPause();
-        mediaPlayer.pause();
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        if(music_switch.isChecked())
-            mediaPlayer.start();
-        else {
-            mediaPlayer.pause();
-        }
     }
 
     @Override
     protected void onStop() {
         super.onStop();
-        //mediaPlayer.stop();
     }
 
     @Override
     protected void onStart() {
         super.onStart();
+        if(SingleGen.getInstance().mediaPlayerIsPlaying())
+            music_switch.setChecked(true);
+        else
+            music_switch.setChecked(false);
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
     }
 }
